==> p0-yelp-json-to-csv-dataset.ipynb <==
import pandas as pd

# My local file path
business_path = r"C:\Users\Rudra\Desktop\yelp\big-data\yelp_academic_dataset_business.json"
checkin_path = r"C:\Users\Rudra\Desktop\yelp\big-data\yelp_academic_dataset_checkin.json"
review_path = r"C:\Users\Rudra\Desktop\yelp\big-data\yelp_academic_dataset_review.json"
tip_path =  r"C:\Users\Rudra\Desktop\yelp\big-data\yelp_academic_dataset_tip.json"
user_path = r"C:\Users\Rudra\Desktop\yelp\big-data\yelp_academic_dataset_user.json"

# I. Business

business_chunks = pd.read_json(business_path, lines=True, chunksize=1000)
business_df = pd.concat(business_chunks, ignore_index=True)

business_df.info()
print(business_df.shape)

# Save to CSV
business_df.to_csv(r"C:\Users\Rudra\Desktop\yelp\data\business.csv", index=False)

business_df.to_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\business.parquet ", index=False)

# II. Tip

tip_chunks = pd.read_json(tip_path, lines=True, chunksize=1000)
tip_df = pd.concat(tip_chunks, ignore_index=True)

tip_df.info()
print(tip_df.shape)

# Save to CSV
tip_df.to_csv(r"C:\Users\Rudra\Desktop\yelp\data\tip.csv", index=False)

tip_df.to_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\tip.parquet ", index=False)

# III. Check In

checkin_chunks = pd.read_json(checkin_path, lines=True, chunksize=1000)
checkin_df = pd.concat(checkin_chunks, ignore_index=True)

checkin_df.info()
print(checkin_df.shape)

# Save to CSV
checkin_df.to_csv(r"C:\Users\Rudra\Desktop\yelp\data\checkin.csv", index=False)

checkin_df.to_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\checkin.parquet ", index=False)

# IV. Review 

review_chunks = pd.read_json(review_path, lines=True, chunksize=1000)
review_df = pd.concat(review_chunks, ignore_index=True)

review_df.info()
print(review_df.shape)

# Save to CSV
review_df.to_csv(r"C:\Users\Rudra\Desktop\yelp\data\review.csv", index=False)

review_df.to_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\review.parquet ", index=False)

# V. User

user_chunks = pd.read_json(user_path, lines=True, chunksize=1000)
user_df = pd.concat(user_chunks, ignore_index=True)

user_df.info()
print(user_df.shape)

# Save to CSV
user_df.to_csv(r"C:\Users\Rudra\Desktop\yelp\data\yelp_user.csv", index=False)

user_df.to_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\yelp_user.parquet ", index=False)
