==> p0-yelp-json-to-parquet-dataset.ipynb <==
import pandas as pd

df = pd.read_csv(r"C:\Users\Rudra\Desktop\yelp\data\business.csv")

df.to_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\business.parquet ", index=False)

df = pd.read_csv(r"C:\Users\Rudra\Desktop\yelp\data\checkin.csv")

df.to_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\checkin.parquet ", index=False)

df = pd.read_csv(r"C:\Users\Rudra\Desktop\yelp\data\review.csv")

df.to_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\review.parquet ", index=False)

df = pd.read_csv(r"C:\Users\Rudra\Desktop\yelp\data\tip.csv")

df.to_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\tip.parquet ", index=False)

df = pd.read_csv(r"C:\Users\Rudra\Desktop\yelp\data\yelp_user.csv")

df.to_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\yelp_user.parquet ", index=False)

df.columns[8]
