==> p3-polars-data-frame-p2.ipynb <==
#!pip install polars -q

import polars as pl
from warnings import filterwarnings

filterwarnings('ignore')

# <strong style="color:#5e17eb">I. Data Frame  </strong>


<strong style="color:#5e17eb"> Parquet </strong>

# df.write_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\business.parquet")

df = pl.DataFrame()
df = pl.read_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\business.parquet")

df.sample()

dff = df.clone()

#  <strong style="color:#5e17eb">  II. DataFrame Attributes & Methods</strong>


## <strong style="color:#5e17eb"> is_duplicated() </strong>


# df[df.is_duplicated()] this is the pandas style
df.filter(df.is_duplicated())

## <strong style="color:#5e17eb"> is_empty() </strong>


df.is_empty()

## <strong style="color:#5e17eb"> extend() </strong>
- When the we merge to data quickly

# df.extend(other_df)

## <strong style="color:#5e17eb">  explode() </strong>


- What: Turns list-like column into multiple rows.

- Modifies Original?  No (returns new df).

- Why: To flatten nested lists.

pl.DataFrame({"a": [[1, 2], [3]]}).explode("a")
# Output: 1 row → 2 rows

## <strong style="color:#5e17eb"> cast() </strong>


df.schema

df.with_columns(
    pl.col("stars").cast(pl.Float32)
).schema

## <strong style="color:#5e17eb">  approx_n_unique()</strong>


df.select(pl.all().approx_n_unique()).collect()

df.approx_n_unique()


## <strong style="color:#5e17eb">  fill_nan() & fill_null()</strong>

| Function      | Fills | Type Targeted | Example Use                    |
| ------------- | ----- | ------------- | ------------------------------ |
| `fill_nan()`  | NaN   | Only float    | Replace invalid floats         |
| `fill_null()` | null  | Any dtype     | Replace `None`, missing values |


| Property | `null`               | `NaN`                    |
| -------- | -------------------- | ------------------------ |
| Type     | Missing value        | Invalid float (IEEE 754) |
| Dtypes   | Any (int, str, etc.) | Only floats              |
| Meaning  | “No value at all”    | “Not a number”           |


## <strong style="color:#5e17eb"> height() & weight() </strong>
- Number of rows

print(df.height)  # Same as df.shape[0]


print(df.width)

## <strong style="color:#5e17eb"> interpolate()  </strong>

- What: Fill missing numeric values linearly.

- Why: For time-series or ordered data gaps.

## <strong style="color:#5e17eb"> hash_rows() </strong>


- What: Create hash values for each row (like fingerprints).

- Why: Fast comparisons, joins, deduplication.

df.hash_rows()

## <strong style="color:#5e17eb"> flags  </strong>


g = pl.DataFrame({
    "col1": [1, 2, 3],
    "col2": ["a", "b", "c"]
})

g.flags


df.flags

## <strong style="color:#5e17eb"> item() </strong>
- Only for series
- What: Return Python scalar from 1-element Series.

pl.Series([42]).item()

## <strong style="color:#5e17eb"> lazy()  </strong>

- What: Convert eager DataFrame to LazyFrame.

- Why: For optimization and deferred execution.

lf = df.lazy()

## <strong style="color:#5e17eb"> limit(n)  </strong>


df.limit(3)

## <strong style="color:#5e17eb"> max() </strong>

max_df_ex = pl.DataFrame({
    "A": [1, 2, 3],
    "B": [6, 7, 8],
    "C": ["apple", "banana", "cherry"],
})

max_values = max_df_ex.max()
print(max_values)


-  max() function on a string column (or a Series of strings), it determines the "maximum" value based on *lexicographical* (dictionary) order. 

df.max()

## <strong style="color:#5e17eb"> max_horizontal()  </strong>
- Row-wise maximum

max_hor_df = pl.DataFrame({
    "col1": [1, 5, 2],
    "col2": [8, 3, 6],
})

row_wise_max = max_hor_df.max_horizontal()
print(row_wise_max)


## <strong style="color:#5e17eb"> map_rows() </strong>
- What: Row-wise Python function application.

- Warning: Slower than expressions. Use only when necessary.

# df.map_rows(lambda row: row[0] + row[1])

## <strong style="color:#5e17eb"> n_unique()  </strong>


# number of unique values
df["stars"].n_unique()

## <strong style="color:#5e17eb"> n_chunks </strong>
- What: Shows number of memory chunks.

- Why: Internal info on memory layout.

- Use: For debugging performance.

df.n_chunks()

print(f"Number of chunks in DataFrame (first series): {df.n_chunks(strategy='first')}")
print(f"Number of chunks in DataFrame (all series): {df.n_chunks(strategy='all')}")

<strong style="color:#5e17eb"> Summary </strong>

- Now, imagine that building the whole castle at once is hard, especially if it's super big, like a giant LEGO city! It might even be too much for your hands to hold all the pieces at once.
- It means that every column in your Polars DataFrame is currently stored in 12 separate chunks or memory allocations
- In summary, the output indicates that your DataFrame's columns are uniformly chunked. This is generally a reasonable state. However, it's a good practice to be aware of the concept of chunking and the availability of `.rechunk()` for potential performance optimization if needed. 

## <strong style="color:#5e17eb"> Pipe </strong>


- It lets you pass a DataFrame through a function as part of a chain.

- Makes your code modular and readable

- Clean chaining of custom Python logic with Polars expressions

def double_stars(df): 
    return df.with_columns((pl.col("stars") * 2).alias("double"))

df.pipe(double_stars)

def extract_categories(df):
    return df.with_columns(
        pl.col("categories")
        .str.split(", ")
        .alias("categories_list")
    )

df = df.pipe(extract_categories)

df["categories_list"]

## <strong style="color:#5e17eb">  Product</strong>


df.select(pl.col("stars").product())


# <strong style="color:#5e17eb"> III. Summary Table </strong>

| Method              | What it Does                           | 
| ------------------- | -------------------------------------- | 
| `extend()`          | Append another df (in-place)           |
| `explode()`         | Flatten list column to rows            | 
| `cast()`            | Change dtype                           |
| `approx_n_unique()` | Fast estimate of unique rows           | 
| `fill_nan()`        | Fill only float NaNs                   | 
| `fill_null()`       | Fill nulls (any type)                  | 
| `fold()`            | Row-wise reduce columns                | 
| `height`            | Row count                              |
| `interpolate()`     | Fill missing values linearly           | 
| `hash_rows()`       | Hash each row                          | 
| `item()`            | Get single value from 1-element Series | 
| `lazy()`            | Convert eager to lazy                  | 
| `limit()`           | First `n` rows                         | 
| `max()`             | Max values per column                  | 
| `map_rows()`        | Row-wise function (slow!)              | 
| `n_unique()`        | Unique value count                     | 
| `n_chunks()`        | Memory layout info                     | 
| `pipe()`            | Chainable functions                    | 
| `product()`         | Multiply all values in a column        | 


<div style="text-align: center;">
  <h4 style="
    display: inline-block;
    color: #5e17eb;
    font-family: 'Segoe UI';
    border-left: 5px solid #729be8ff;
    background-color: #F8F9F9;
    padding: 10px 20px;
    border-radius: 5px;
    text-align: left;
  "><b>
    Thank You 💜
    </b>
  </h4>
</div>
