==> p4-selecting-filtering-editing.ipynb <==
import polars as pl
from warnings import filterwarnings

filterwarnings('ignore')

<strong style="color:#5e17eb"> Parquet </strong>
- Polars + Parquet == More speed
- So we converted all json -> csv -> parquet file 

df = pl.read_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\business.parquet")

df.sample()

user_df = pl.scan_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\yelp_user.parquet")

# <strong style="color:#5e17eb"> I. Selecting Data </strong>


## <strong style="color:#5e17eb">  Column Selection </strong>


df.select([
    'business_id',
    'name',
    'address',
    'city',
    'state'
]).sample(2)

df[['stars','review_count','is_open','attributes']].sample(1)

## <strong style="color:#5e17eb">  Row Selection </strong>


df[55:58]

# Select the row start from 12 and take 3 rows
# slice(where to start, how many to select)
df.slice(12,3)

## <strong style="color:#5e17eb">  Cell Selection </strong>

- Row + Column

df[4:9:2, ['name', 'city']]

## <strong style="color:#5e17eb"> Boolean Row Filtering </strong>


df.filter(
    (pl.col("stars") > 4.2),
    (pl.col("city") == "Green Lane"),
    (pl.col("is_open") == 1)
)

## <strong style="color:#5e17eb"> Conditional Selection </strong>


df.filter(
    (pl.col("stars") > 4.2) &
    (pl.col("city") == "Green Lane")
)

## <strong style="color:#5e17eb"> Expression based Selection </strong>


df.select(
   (pl.col("stars") + 10).alias("star_plus_ten")
)

## <strong style="color:#5e17eb"> Summary </strong>



**Goal: View only relevant parts of your data**

| **Sub-Topic**                    | **Description / Use**          | **Example**                                                  |
| -------------------------------- | ------------------------------ | ------------------------------------------------------------ |
| **1.1 Column Selection**         | Select one or multiple columns | `df.select(["name", "price"])`                               |
| **1.2 Row Selection (by index)** | Select rows using slicing      | `df[0:5]` or `df.slice(0, 5)`                                |
| **1.3 Cell Selection**           | Select row + column            | `df[3, "name"]`                                              |
| **1.4 Boolean Row Filtering**    | Select rows by condition       | `df.filter(pl.col("price") > 100)`                           |
| **1.5 Conditional Selection**    | Multiple conditions            | `df.filter((pl.col("price") > 50) & (pl.col("rating") > 4))` |
| **1.6 Select with `pl.col()`**   | Expression-based selection     | `df.select(pl.col("price") + 10)`                            |


# <strong style="color:#5e17eb"> II. Editing / Modifying the Data Frame </strong>


## <strong style="color:#5e17eb"> Add New Column </strong>


df.with_columns(
    (pl.col("review_count") + (pl.col("stars"))).alias("review_count_plus_stars") 
).select(["review_count_plus_stars"])

# Multiple columns
df.with_columns([
    (pl.col("stars") * 2).alias("double_stars"),
    (pl.col("name").str.len_chars().alias("len_name"))
])[["double_stars", "len_name"]]

# single columns
df.with_columns(
     (pl.col("stars") * 2).alias("double_stars")
)["double_stars"]

## <strong style="color:#5e17eb">  Modify Column Values	 </strong>


df.with_columns(
    (pl.col("stars") + 2).alias("stars_2")
)['stars_2']

## <strong style="color:#5e17eb">Rename Columns  </strong>


df.rename({ "hours" : "week_hours"}).sample()

## <strong style="color:#5e17eb">   Drop Columns</strong>


df.drop(["postal_code"]).sample()

## <strong style="color:#5e17eb">  Change Data Types </strong>


df.with_columns(
    (pl.col("stars")).cast(pl.Int32)
).schema

## <strong style="color:#5e17eb">  Fill Missing Values</strong>


df.null_count()

# fill with mean values
mean_val = df.select(pl.col("stars").mean()).item()
mean_val

df.with_columns(
    (pl.col("stars").fill_null(mean_val)).alias("stars_mean")
)['stars_mean'].sample(3)

df.with_columns(
    pl.col("stars").fill_nan(0).alias("stars_no_nan")
)['stars_no_nan'].sample(3)

# Null values fill with the unknown
df = df.with_columns(
    (pl.col("address").fill_null("Unknown")).alias("fill_address")
)

df.filter(pl.col("fill_address") == "Unknown").shape

| Fill Method           | Code Example                                             |
| --------------------- | -------------------------------------------------------- |
| Constant              | `fill_null(0)` or `fill_null("Unknown")`                 |
| Mean                  | `fill_null(df["col"].mean())` (use `.item()` to extract) |
| Forward Fill (ffill)  | `fill_null(strategy="forward")`                          |
| Backward Fill (bfill) | `fill_null(strategy="backward")`                         |
| Fill NaN (not null)   | `fill_nan(0)`                                            |


## <strong style="color:#5e17eb">   Replace Values</strong>


# Unknown values replace with the missing
df = df.with_columns(
    (pl.col("fill_address").replace({"Unknown" : "Missing"}))
)

df.filter(pl.col("fill_address") == "Missing").shape

## <strong style="color:#5e17eb"> Round / Format Columns  </strong>


df.with_columns(
    (pl.col("stars").round())
)['stars'].sample(5)

## <strong style="color:#5e17eb"> Summary </strong>



**Goal: Change or enrich the structure/content**


| **Sub-Topic**                  | **Description / Use**              | **Example**                                                 |
| ------------------------------ | ---------------------------------- | ----------------------------------------------------------- |
| **2.1 Add New Column**         | Add derived column                 | `df.with_columns((pl.col("a") + pl.col("b")).alias("sum"))` |
| **2.2 Modify Column Values**   | Change existing column             | `df.with_columns(pl.col("price") * 2)`                      |
| **2.3 Rename Columns**         | Rename one or more columns         | `df.rename({"old": "new"})`                                 |
| **2.4 Drop Columns**           | Remove unwanted columns            | `df.drop(["extra", "useless"])`                             |
| **2.5 Change Data Types**      | Cast data to specific types        | `df.with_columns(pl.col("age").cast(pl.Int32))`             |
| **2.6 Fill Missing Values**    | Fill `null` with value or strategy | `df.fill_null(0)`                                           |
| **2.7 Replace Values**         | Replace specific values            | `df.with_columns(pl.col("status").replace("Old", "New"))`   |
| **2.8 Round / Format Columns** | Round floats or format strings     | `pl.col("score").round(2)`                                  |


# <strong style="color:#5e17eb"> III. Rearranging / Reordering </strong>


## <strong style="color:#5e17eb"> Reorder Columns  </strong>


df.select(["business_id", "name", "city"])

## <strong style="color:#5e17eb">  Sort Rows</strong>


# sort the values in descending orders
df  = df.sort("stars", descending=True)

df['stars']

## <strong style="color:#5e17eb">Reverse Rows  </strong>


# Reverse the rows
df['stars'].reverse()

## <strong style="color:#5e17eb">Drop Duplicate Rows  </strong>


print("Unique postal code is :",df['postal_code'].unique().shape[0])

## <strong style="color:#5e17eb"> Limit Rows </strong>

- slice()
- head()
- tail()
- sample()

df.slice(2, 3)

## <strong style="color:#5e17eb"> Summary </strong>

 **Goal: Change how data is organized**

| **Sub-Topic**               | **Description / Use**         | **Example**                          |
| --------------------------- | ----------------------------- | ------------------------------------ |
| **3.1 Reorder Columns**     | Move column order             | `df.select(["id", "name", "price"])` |
| **3.2 Sort Rows**           | Sort rows by column           | `df.sort("rating", descending=True)` |
| **3.3 Reverse Rows**        | Flip the row order            | `df.reverse()`                       |
| **3.4 Drop Duplicate Rows** | Remove repeated rows          | `df.unique()`                        |
| **3.5 Limit Rows**          | Select top-N or bottom-N rows | `df.head(10)`, `df.tail(5)`          |


# <strong style="color:#5e17eb"> IV. Conditional Logic & Feature Engineering  </strong>


## <strong style="color:#5e17eb">   Conditional Columns</strong>


df = df.with_columns([
    pl.when(pl.col("stars") >= 4.5).then(pl.lit("Excellent"))\
    .when(pl.col("stars") >= 3.5).then(pl.lit("Good"))\
    .when(pl.col("stars") >= 2.5).then(pl.lit("Average"))\
    .otherwise(pl.lit("Poor"))\
    .alias("rating_label")
])

df.group_by("rating_label").agg(pl.count())

`pl.lit()` is a fundamental tool for incorporating static, constant values into dynamic Polars expressions, enabling flexible and efficient data manipulations.

## <strong style="color:#5e17eb">  Binning / Bucketing</strong>


df = df.with_columns([
    pl.when((pl.col("is_open") == 1) & (pl.col("review_count") > 100))
    .then(pl.lit("Active"))
    .otherwise(pl.lit("Inactive"))
    .alias("activity_status")
])

df.group_by("activity_status").agg(pl.count())

| Goal                        | Method Used                     |
| --------------------------- | ------------------------------- |
| Categorize values           | `pl.when().then().otherwise()`  |
| Element-wise Python logic   | `pl.col().map_elements()`       |
| Combine fields into strings | `pl.struct(...).map_elements()` |


## <strong style="color:#5e17eb">  Map / Custom Functions</strong>


df = df.with_columns(
    pl.col("review_count").map_elements(
        lambda x: "Popular" if x > 500 else "Regular"
    ).alias("popularity")
)

df.group_by("popularity").agg(pl.count())

df = df.with_columns([
    pl.when(pl.col("longitude") < -100).then(pl.lit("West"))
    .when(pl.col("longitude") > -80).then(pl.lit("East"))
    .otherwise(pl.lit("Central"))
    .alias("region")
])

df.select(["name", "longitude", "region"])

## <strong style="color:#5e17eb"> Summary </strong>

 **Goal: Build intelligent features or rules**
| **Sub-Topic**                  | **Description / Use**                  | **Example**                                      |
| ------------------------------ | -------------------------------------- | ------------------------------------------------ |
| **4.1 Conditional Columns**    | Create values based on `if/else` logic | `pl.when(...).then(...).otherwise(...)`          |
| **4.2 Binning / Bucketing**    | Create bins from continuous data       | `pl.when(pl.col("age") > 50).then("Old")`        |
| **4.3 Map / Custom Functions** | Apply custom logic                     | `df.with_columns(pl.col("x").map_elements(...))` |


 # <strong style="color:#5e17eb"> V.  Inspecting & Validating
 </strong>


## <strong style="color:#5e17eb"> View Schema </strong>


df.schema

## <strong style="color:#5e17eb"> Summary Stats </strong>


df.describe()

## <strong style="color:#5e17eb">  Check Column Nulls </strong>


df.null_count()

`df.iter_columns()`
- Iterates over Series (each column) in the DataFrame.

for col in df.iter_columns():
    print(col.name, col.dtype, col[1:4])

| Function                  | What it Does                       | Returns            | Use Case                         |
| ------------------------- | ---------------------------------- | ------------------ | -------------------------------- |
| `with_column()`           | Add/modify one column              | New DataFrame      | Add a new computed field         |
| `with_columns([...])`     | Add/modify multiple columns        | New DataFrame      | Add several fields efficiently   |
| `iter_columns()`          | Iterate over Series (columns)      | Iterator of Series | Inspect or loop through columns  |
| `iter_columns_seq()`      | Indexable list of Series (columns) | Sequence of Series | Use indexing (`cols[0]`, etc.)   |
| `get_column("col")`       | Get a single column as Series      | `pl.Series`        | Direct column access (read-only) |
| `find_idx_by_name("col")` | Get column index                   | Integer            | Map name → position in DataFrame |


## <strong style="color:#5e17eb">  Unique / Value Counts</strong>


df['attributes'].unique_counts()

## <strong style="color:#5e17eb"> Summary </strong>
**Goal: Check the structure and correctness**

| **Sub-Topic**                 | **Description / Use**           | **Example**                              |
| ----------------------------- | ------------------------------- | ---------------------------------------- |
| **5.1 View Schema**           | See column names and types      | `df.schema`                              |
| **5.2 Summary Stats**         | Quick stats for numeric columns | `df.describe()`                          |
| **5.3 Check Column Nulls**    | Check for missing values        | `df.null_count()`                        |
| **5.4 Unique / Value Counts** | Count distinct values           | `df.select(pl.col("category").unique())` |


<div style="text-align: center;">
  <h4 style="
    display: inline-block;
    color: #5e17eb;
    font-family: 'Segoe UI';
    border-left: 5px solid #5e17eb;
    background-color: #F8F9F9;
    padding: 10px 20px;
    border-radius: 5px;
    text-align: left;
  ">
  <b>
    Thank You 💜
    </b>
  </h4>
</div>
