==> p5-group-by.ipynb <==
import polars as pl
from warnings import filterwarnings

filterwarnings('ignore')

df = pl.read_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\business.parquet")

user_lf = pl.scan_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\yelp_user.parquet")

df.sample(5)

# <strong style="color:#5e17eb"> 1. Common group_by Aggregations </strong>


df.with_columns([
    (pl.col("stars").mean()).alias("mean_starts"),
    (pl.col("stars").median()).alias("median_starts"),
])[['mean_starts', 'median_starts']].item

df.select(pl.col("stars").sum()).item()

df.select(pl.col("stars").n_unique())

df.select(pl.col("stars").unique())

## <strong style="color:#5e17eb"> Summary </strong>


**Common group_by Aggregations**


* `pl.count()` → count of rows
* `pl.col("col").mean()` → average
* `pl.col("col").sum()` → sum
* `pl.col("col").n_unique()` → unique count
* `pl.col("col").min()`, `pl.col("col").max()` → min/max
* `pl.col("col").first()`, `pl.col("col").last()` → first/last row
* `pl.col("col").list()` → all values in a list



# <strong style="color:#5e17eb"> 2. Basic group_by </strong>


## <strong style="color:#5e17eb">  Single Column </strong>

Find out mean of stars for each city

df.group_by("city").agg(
    (pl.col("stars").mean()).alias("stars_mean"),
)

## <strong style="color:#5e17eb">  Multiple Columns </strong>

df.group_by(["state", "city"]).agg(
    pl.col("stars").mean()
)

## <strong style="color:#5e17eb">Aggregation Functions  </strong>
- Mean, sum, count, min, max, median, etc.

df.filter(
    pl.col("state") == "PA"
).shape

Find out city on the NJ state, with average stars and total review of that city?

df.filter(pl.col("state") == "NJ") \
    .group_by(["state", "city"]) \
    .agg(
        pl.col("stars").mean().alias("avg_stars"),
        pl.col("review_count").sum().alias("total_review")
    )

## <strong style="color:#5e17eb">  Multiple Aggregations</strong>

df.group_by("state").agg(
    pl.col("stars").mean().alias("mean_stars"),
    pl.col("stars").std().alias("std_stars"),
)

## <strong style="color:#5e17eb">  Multiple Aggregations columns </strong>

df.group_by("state").agg(pl.all().mean())['state','stars', 'review_count']

## <strong style="color:#5e17eb"> Summary </strong>


 **Goal: Aggregate data by one or more columns**
 
 | **Sub-Topic**                        | **Description**                          | **Example (Yelp)**                                                         |
| ------------------------------------ | ---------------------------------------- | -------------------------------------------------------------------------- |
| **Single Column**                | Group by one column and summarize        | `df.group_by("city").agg(pl.col("stars").mean())`                           |
| **Multiple Columns**             | Group by more than one column            | `df.group_by(["state", "city"]).agg(pl.col("stars").mean())`                |
| **Aggregation Functions**        | Mean, sum, count, min, max, median, etc. | `df.group_by("state").agg(pl.col("review_count").sum())`                    |
| **Multiple Aggregations**        | Apply multiple functions to same col     | `df.group_by("state").agg([pl.col("stars").mean(), pl.col("stars").max()])` |
| **Aggregating Multiple Columns** | Summary across multiple columns          | `df.group_by("state").agg(pl.all().mean())`                                 |


# <strong style="color:#5e17eb"> 3. group_by with Expressions </strong>


## <strong style="color:#5e17eb">Derived Columns in Group  </strong>


# No null value in the review column
df.select(pl.col("review_count").is_null().sum()).item()

df['review_count'].describe()

total_reviews = df["review_count"].sum()

df.group_by("city") \
    .agg([
        # Share of reviews (rounded to 2 decimals)
        (pl.col("review_count").sum() / total_reviews).round(2).alias("share"),

        # Average stars per city
        pl.col("stars").mean().round(2).alias("avg_stars"),

        # Average reviews per business in the city
        (pl.col("review_count").mean().round(1)).alias("review_per_business")
    ])


## <strong style="color:#5e17eb"> Conditional Aggregation </strong>


df.group_by("state").agg(
    pl.col("stars")
        .filter(pl.col("is_open") == 1)
        .mean()
        .alias("open_avg_stars")
)

# 2nd way to do this
df.filter(pl.col("is_open") == 1) \
    .group_by("state") \
    .agg(
        pl.col("stars").mean().alias("open_avg_stars")
    )

## <strong style="color:#5e17eb"> Aggregating Unique Values </strong>


df.group_by("state").agg(
    pl.col("city").n_unique().alias("unique_cities")
)

## <strong style="color:#5e17eb"> Summary </strong>


**Goal: Use transformations inside group_by**

| **Sub-Topic**                     | **Description**                          | **Example (Yelp)**                                                                                       |
| --------------------------------- | ---------------------------------------- | -------------------------------------------------------------------------------------------------------- |
| **Derived Columns in Group**  | Create calculated metrics inside group_by | `df.group_by("city").agg((pl.col("review_count") / pl.col("review_count").sum()).alias("share"))`         |
| **Conditional Aggregation**   | Aggregate only subset of data            | `df.group_by("state").agg(pl.col("stars").filter(pl.col("is_open") == 1).mean().alias("open_avg_stars"))` |
| **Aggregating Unique Values** | Count distinct entries                   | `df.group_by("state").agg(pl.col("city").n_unique().alias("unique_cities"))`                              |


# <strong style="color:#5e17eb">  4. group_by with `group_by_dynamic`</strong>

- In this notebook we only focused on the *group_by_dynamic* works 
- In this **Polars Series** we discussed the date & time properly
- Before applying the functions always sort the values

- https://docs.pola.rs/api/python/stable/reference/dataframe/api/polars.DataFrame.group_by_dynamic.html



## <strong style="color:#5e17eb">  Date/Time Based </strong>

- For data and time we used yelp_user dataset
- First casting the data type

user_lf.select("yelping_since").limit(1).collect()

user_lf = user_lf.with_columns(
    pl.col("yelping_since").str.strptime(pl.Datetime, format="%Y-%m-%d %H:%M:%S")
)

user_lf.select("yelping_since").limit(3).collect()

user_lf \
    .sort("yelping_since") \
    .group_by_dynamic(
        index_column="yelping_since",
        every='1y'
    ).agg(
        pl.col("review_count").sum().alias("total_reviews"),
        pl.count().alias("user_count")
    ).collect()

## <strong style="color:#5e17eb"> Rolling Windows </strong>


user_lf \
    .sort("yelping_since") \
    .group_by_dynamic(
        index_column="yelping_since",
        every='6mo',
        period='2y',
    ).agg([
        pl.col("review_count").sum().alias("rolling_total_review"),
        pl.count().alias("user_in_window_of_6mo"),
    ]).collect()

## <strong style="color:#5e17eb"> Summary </strong>


**Goal: Time-based or range-based grouping**


| **Sub-Topic**               | **Description**                             | **Example (Yelp)**                                                              |
| --------------------------- | ------------------------------------------- | ------------------------------------------------------------------------------- |
| **Date/Time Based**     | Aggregate per fixed time windows            | `df.group_by_dynamic("date_col", every="1mo").agg(pl.col("review_count").sum())` |
| **Rolling Windows**     | Use rolling intervals                       | `df.group_by_dynamic("date_col", every="7d", period="7d").agg(...)`              |



# <strong style="color:#5e17eb"> 5. group_by Advanced Methods  </strong>


## <strong style="color:#5e17eb">  `.over()` Window Functions</strong>


df['state'].n_unique()

df.with_columns(
    pl.col("stars").mean().over("state").alias("state_avg")
)['state','stars', 'state_avg']

## <strong style="color:#5e17eb"> Sorting Inside Groups </strong>


df.group_by("state").agg(
    pl.col("review_count").mean().alias("avg_review_count")
).sort("avg_review_count", descending=True)  


## <strong style="color:#5e17eb"> Keep First/Last in Groups </strong>


df.sort("state") \
    .group_by("state").agg(
        pl.col("stars").mean()
    ).limit(4)

## <strong style="color:#5e17eb">  List Aggregation </strong>


df.group_by("state").agg(
    pl.col("city").alias("city_list")
)

## <strong style="color:#5e17eb"> Summary </strong>


 **Goal: Go beyond simple aggregation**

| **Sub-Topic**                      | **Description**                         | **Example (Yelp)**                                                         |
| ---------------------------------- | --------------------------------------- | -------------------------------------------------------------------------- |
| **`.over()` Window Functions** | Running stats without collapsing groups | `df.with_columns(pl.col("stars").mean().over("state").alias("state_avg"))` |
| **Sorting Inside Groups**      | Sort by a metric within groups          | `df.group_by("state").agg(pl.col("review_count").sort(descending=True))`    |
| **Keep First/Last in Groups**  | Top/bottom rows in each group           | `df.group_by("state").agg(pl.col("stars").first())`                         |
| **List Aggregation**           | Keep group data as lists                | `df.group_by("state").agg(pl.col("city").alias("cities_list"))`             |


# <strong style="color:#5e17eb"> 6. Looping over Groups </strong>


## <strong style="color:#5e17eb"> Convert to Pandas Loop </strong>


for group_key, group_df in df.group_by("state"):
    print(group_key, group_df.limit(1))

## <strong style="color:#5e17eb"> Process in Lazy Mode </strong>

- Use `.collect()` during the lazy mode

user_lf.head(2).collect()

## <strong style="color:#5e17eb">  Apply Functions to Each Group</strong>


### <strong style="color:#5e17eb"> A. Using map_groups() </strong>


# Custom logic: Weighted average stars per city
df.group_by("city") \
    .map_groups(
        lambda group_df: pl.DataFrame(
            {
                "city" : [group_df["city"][0]],
                "weighted_avg_stars" : [
                    (group_df["stars"] * group_df["review_count"]).sum() / group_df["review_count"].sum()
                ]
            }
        )
    )

### <strong style="color:#5e17eb">  Multiple custom calculations</strong>


df.group_by("state") \
    .map_groups(
        lambda g : pl.DataFrame(
            {
                "state" : [g["state"][0]],
                "avg_stars" : [g["stars"].mean()],
                "popularity_score" : [g["review_count"].max() * g["stars"].mean()],
            }
        )
    )

### <strong style="color:#5e17eb"> Conditional inside group </strong>


df.filter(pl.col("is_open") == 1) \
.group_by("state") \
    .map_groups(
        lambda g : pl.DataFrame(
            {
                "state" : [g["state"][0]],
                "open_avg_stars" : [g["stars"].mean()],
            }
        )
    )

## <strong style="color:#5e17eb"> Summary </strong>


**Goal: Process each group separately**


| **Sub-Topic**                         | **Description**                        | **Example (Yelp)**                                                                          |
| ------------------------------------- | -------------------------------------- | ------------------------------------------------------------------------------------------- |
| **Convert to Pandas Loop**        | Use `df.group_by()` and iterate (eager) | `for group_key, group_df in df.group_by("state"):`                                           |
| **Process in Lazy Mode**          | Use `.collect()` after filtering       | `for state in df.select("state").unique().to_series(): df.filter(pl.col("state") == state)` |
| **Apply Functions to Each Group** | Run custom function over groups        | Custom Python function with `.map_groups()` *(in Rust backend)*                             |


<div style="text-align: center;">
  <h4 style="
    display: inline-block;
    color: #5e17eb;
    font-family: 'Segoe UI';
    border-left: 5px solid #5e17eb;
    background-color: #F8F9F9;
    padding: 10px 20px;
    border-radius: 5px;
    text-align: left;
  ">
  <b>
    Thank You 💜
    </b>
  </h4>
</div>
