==> p7-polar-date-and-time.ipynb <==
import polars as pl
import datetime
from warnings import filterwarnings

filterwarnings('ignore')

business_lf = pl.scan_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\business.parquet")
checkin_lf = pl.scan_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\checkin.parquet")
review_lf = pl.scan_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\review.parquet")
tip_lf = pl.scan_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\tip.parquet")
user_lf = pl.scan_parquet(r"C:\Users\Rudra\Desktop\yelp\parquet-data\yelp_user.parquet")

user_lf.collect_schema()

user_lf.select(['yelping_since']).collect()

# here is code

user_lf = pl.DataFrame({
    "user_id": ["u1", "u2", "u3"],
    "yelping_since": [
        "2007-01-25 16:47:26",
        "2009-01-25 04:35:42",
        "2020-12-19 02:32:39"
    ]
}).lazy()


# <strong style="color:#5e17eb"> 1. Creating & Parsing Date/Time Columns </strong>


## <strong style="color:#5e17eb"> Parse from String </strong>


user_lf.head(1).collect()

%Y = year, %m = month, %d = day, %H = hour, %M = minute, %S = seconds.



user_lf.with_columns(
    pl.col("yelping_since") \
        .str.strptime(pl.Datetime, "%Y-%m-%d %H:%M:%S")
        .alias("yelping_since_dt")
).select(['yelping_since',	'yelping_since_dt']).collect()

## <strong style="color:#5e17eb">Auto-parse  </strong>


user_lf.with_columns(
    pl.col("yelping_since") 
        .str.to_datetime()
        .alias("yelping_since_dt")
).select(['yelping_since',	'yelping_since_dt']).collect()

## <strong style="color:#5e17eb"> From Timestamps </strong>


pl.DataFrame({
    "custom_date": [pl.date(2025, 8, 5)]
})


## <strong style="color:#5e17eb">Create from Components  </strong>


user_lf.with_columns(
    pl.col("yelping_since").str.to_datetime().alias("yelping_since_dt")
).with_columns([
    pl.col("yelping_since_dt").dt.year().alias("year_joined"),
    pl.col("yelping_since_dt").dt.month().alias("month_joined"),
    pl.col("yelping_since_dt").dt.day().alias("days_joined"),
    (pl.lit("2025-08-06").str.to_datetime() - pl.col("yelping_since_dt")).alias("membership_duration")
]).collect()

## <strong style="color:#5e17eb">  From Epoch Seconds</strong>


user_lf_parsed = user_lf.with_columns(
    pl.col("yelping_since").str.strptime(pl.Datetime, format="%Y-%m-%d %H:%M:%S").alias("yelping_dt")
)
user_lf_parsed.select(["yelping_since", "yelping_dt"]).collect().head()

user_lf = user_lf.with_columns(
    pl.col("yelping_since").str.to_datetime().alias("yelping_dt")
)["yelping_dt"].collect()


user_lf_epoch = user_lf.with_columns(
    pl.col("yelping_since")
      .str.strptime(pl.Datetime, "%Y-%m-%d %H:%M:%S")
      .dt.epoch("d")
      .alias("epoch_days")
)
user_lf_epoch.collect()

user_lf.select([
    pl.col("yelping_since").str.to_datetime().dt.year().alias("year"),
    pl.col("yelping_since").str.to_datetime().dt.month().alias("month"),
    pl.col("yelping_since").str.to_datetime().dt.weekday().alias("weekday")
]).collect()


user_lf.with_columns(
    pl.col("yelping_since").str.to_datetime().alias("yelping_dt")
).collect()

## <strong style="color:#5e17eb"> Summary </strong>

**Goal: Convert raw strings or numbers into Polars date/time types.**
| Sub-Topic                      | Function / Method                                        | Example                                                |
| ------------------------------ | -------------------------------------------------------- | ------------------------------------------------------ |
| **1.1 Parse from String**      | `.str.strptime(dtype, format=...)`                       | `pl.col("date_str").str.strptime(pl.Date, "%Y-%m-%d")` |
| **1.2 Auto-parse**             | `.str.to_date()`, `.str.to_datetime()`, `.str.to_time()` | `pl.col("date_str").str.to_datetime()`                 |
| **1.3 From Timestamps**        | `.cast(pl.Datetime)`                                     | `pl.col("timestamp").cast(pl.Datetime)`                |
| **1.4 Create from Components** | `pl.date(year, month, day)`                              | `pl.date(2025, 8, 5)`                                  |
| **1.5 From Epoch Seconds**     | `.dt.epoch(time_unit="s")`                               | `pl.Series([1620000000]).cast(pl.Datetime("s"))`       |



# <strong style="color:#5e17eb"> 2. Extracting Date/Time Parts </strong>


## <strong style="color:#5e17eb">Calender Parts  </strong>

user_lf.select([
    pl.col("yelping_dt").dt.year().alias("year"),
    pl.col("yelping_dt").dt.month().alias("month"),
    pl.col("yelping_dt").dt.day().alias("day")
]).collect()

## <strong style="color:#5e17eb">  Time Pats</strong>

calendar_parts = user_lf.with_columns([
    pl.col("yelping_dt").dt.year().alias("year"),
    pl.col("yelping_dt").dt.month().alias("month"),
    pl.col("yelping_dt").dt.day().alias("day")
])

calendar_parts.select(["yelping_since", "year", "month", "day"]).collect()


## <strong style="color:#5e17eb"> Week and day Info
 </strong>

time_parts = user_lf.with_columns([
    pl.col("yelping_dt").dt.hour().alias("hour"),
    pl.col("yelping_dt").dt.minute().alias("minute"),
    pl.col("yelping_dt").dt.second().alias("second"),
    pl.col("yelping_dt").dt.millisecond().alias("millisecond")
])

time_parts.select(["yelping_since", "hour", "minute", "second", "millisecond"]).collect().head()


week_day_info = user_lf.with_columns([
    pl.col("yelping_dt").dt.week().alias("week"),
    pl.col("yelping_dt").dt.weekday().alias("weekday"),       # Monday=0
    pl.col("yelping_dt").dt.ordinal_day().alias("ordinal_day") # Day of year (1-366)
])

week_day_info.select(["yelping_since", "week", "weekday", "ordinal_day"]).collect().head()


## <strong style="color:#5e17eb"> Start/end Info  </strong>

start_end_period = user_lf.with_columns([
    # Start of month/year
    pl.col("yelping_dt").dt.truncate("1mo").alias("start_of_month"),
    pl.col("yelping_dt").dt.truncate("1y").alias("start_of_year"),

    # End of month/year
    (pl.col("yelping_dt").dt.truncate("1mo") + pl.duration(days=31)).dt.truncate("1mo").alias("end_of_month"),
    (pl.col("yelping_dt").dt.truncate("1y") + pl.duration(days=366)).dt.truncate("1y").alias("end_of_year")
])

start_end_period.select([
    "yelping_dt", "start_of_month", "end_of_month", "start_of_year", "end_of_year"
]).collect()

## <strong style="color:#5e17eb"> Summary </strong>

**Goal: Break date/time into components for analysis.**

| Sub-Topic                   | Function / Method                                                 | Example                               |
| --------------------------- | ----------------------------------------------------------------- | ------------------------------------- |
| **2.1 Calendar Parts**      | `.dt.year()`, `.dt.month()`, `.dt.day()`                          | `df.select(pl.col("date").dt.year())` |
| **2.2 Time Parts**          | `.dt.hour()`, `.dt.minute()`, `.dt.second()`, `.dt.millisecond()` | `df.select(pl.col("time").dt.hour())` |
| **2.3 Week & Day Info**     | `.dt.week()`, `.dt.weekday()`, `.dt.ordinal_day()`                | `pl.col("date").dt.weekday()`         |
| **2.4 Start/End of Period** | `.dt.start_of(unit)`, `.dt.end_of(unit)`                          | `pl.col("date").dt.start_of("month")` |


# <strong style="color:#5e17eb">  3. Manipulating Dates & Times</strong>


> Offsets support d (days), h (hours), m (minutes), s (seconds), mo (months), y (years).

## <strong style="color:#5e17eb">  Add & subtract </strong>

add_subtract = user_lf.with_columns([
    pl.col("yelping_dt").dt.offset_by("1d").alias("plus_1_day"),
    pl.col("yelping_dt").dt.offset_by("-2h").alias("minus_2_hours")
])

add_subtract.select(["yelping_dt", "plus_1_day", "minus_2_hours"]).collect().head()


## <strong style="color:#5e17eb">Rounding  </strong>

rounding = user_lf.with_columns([
    pl.col("yelping_dt").dt.round("1h").alias("round_to_hour"),
    pl.col("yelping_dt").dt.truncate("1d").alias("truncate_to_day"),
    pl.col("yelping_dt").dt.truncate("1w").alias("truncate_to_week")
])
rounding.select(["yelping_dt", "round_to_hour", "truncate_to_day", "truncate_to_week"]).collect().head()

## <strong style="color:#5e17eb">  Replace Components </strong>

replace_parts = user_lf.with_columns(
    pl.col("yelping_dt").dt.replace(year=2025, month=12).alias("replaced_date")
)

replace_parts.select(["yelping_since", "replaced_date"]).collect().head()


## <strong style="color:#5e17eb">Clips Ranges  </strong>

clip_ranges = user_lf.with_columns(
    pl.col("yelping_dt").clip(
        pl.date(2007, 1, 1), 
        pl.date(2010, 1, 1)
    ).alias("clipped_date")
)

clip_ranges.select(["yelping_since", "clipped_date"]).collect().head()


## <strong style="color:#5e17eb"> Summary </strong>

**Goal: Shift, round, and adjust date/time values.**

| Sub-Topic                  | Function / Method                                      | Example                                |
| -------------------------- | ------------------------------------------------------ | -------------------------------------- |
| **3.1 Add/Subtract**       | `.dt.offset_by("1d")`, `.dt.offset_by("-2h")`          | `pl.col("date").dt.offset_by("7d")`    |
| **3.2 Rounding**           | `.dt.round("1h")`, `.dt.floor("1d")`, `.dt.ceil("1w")` | `pl.col("date").dt.round("1h")`        |
| **3.3 Replace Components** | `.dt.replace(year=2025, month=12)`                     | `pl.col("date").dt.replace(year=2025)` |
| **3.4 Clip Ranges**        | `.clip(pl.date(2024,1,1), pl.date(2025,1,1))`          | Restrict values to range               |


# <strong style="color:#5e17eb"> 4. Time Differences & Durations </strong>


## <strong style="color:#5e17eb"> total_# </strong>

user_lf.with_columns(
    pl.col("yelping_dt")
      .diff(n=1, null_behavior="ignore")      
      .dt.total_days()                         
      .alias("diff_in_days")
).select(['yelping_dt', 'diff_in_days']).collect()


## <strong style="color:#5e17eb"> diff </strong>

user_lf.with_columns(
    pl.col("yelping_dt") \
        .diff().dt.total_days().alias("days_between")
).select(["yelping_dt", "days_between"]).collect()

## <strong style="color:#5e17eb"> Summary </strong>


**Goal: Measure time gaps and durations.**

| Sub-Topic                   | Function / Method                              | Example                           |
| --------------------------- | ---------------------------------------------- | --------------------------------- |
| **4.1 Difference in Units** | `.dt.diff()`                                   | `pl.col("date").diff().dt.days()` |
| **4.2 Duration Conversion** | `.dt.total_days()`, `.dt.total_seconds()`      | `pl.col("end") - pl.col("start")` |
| **4.3 Absolute Time Gaps**  | `(pl.col("end") - pl.col("start")).dt.hours()` | Difference in hours               |


# <strong style="color:#5e17eb"> 5. Grouping & Resampling with Dates </strong>


## <strong style="color:#5e17eb"> Regular Group_by </strong>

user_lf = user_lf.with_columns(
    pl.col("yelping_since").str.strptime(pl.Datetime)
)

user_lf.head(1).collect()

## <strong style="color:#5e17eb"> Time based Grouping </strong>

business_lf.group_by("review_count") \
    .agg([
        (pl.col("stars").mean())
    ]).collect()

## <strong style="color:#5e17eb"> Rolling Windows </strong>

user_lf.sort("yelping_since") \
    .group_by_dynamic(
        index_column="yelping_since",  # Must be datetime
        every="1y",
        period="2y"
    ).agg([
        pl.col("review_count").mean(),
        pl.count().alias("review number")
    ]).collect()


## <strong style="color:#5e17eb"> Summary </strong>

**Goal: Aggregate by date/time buckets.**

| Sub-Topic                   | Function / Method                                   | Example             |
| --------------------------- | --------------------------------------------------- | ------------------- |
| **5.1 Regular group_by**     | `.group_by("date").agg(...)`                         | Group by exact date |
| **5.2 Time-based Grouping** | `.group_by_dynamic("date", every="1mo")`             | Monthly aggregation |
| **5.3 Rolling Windows**     | `.group_by_dynamic("date", every="7d", period="7d")` | Weekly rolling sums |


# <strong style="color:#5e17eb"> 6. Working with Timezones </strong>


df = pl.DataFrame({
    "date_str": ["2024-01-01", "2024-06-15", "2024-12-31"]
})

# Convert to Date
df = df.with_columns(
    pl.col("date_str").str.strptime(pl.Date, "%Y-%m-%d").alias("date")
)
print(df)


## <strong style="color:#5e17eb"> Set a Time Zone </strong>

df = pl.DataFrame({
    "date_slash": ["01/02/2024", "15/03/2024"]
})

df = df.with_columns(
    pl.col("date_slash").str.strptime(pl.Date, "%d/%m/%Y").alias("date")
)


df = pl.DataFrame({
    "ts": [
        datetime.datetime(2024, 1, 1, 12, 0),
        datetime.datetime(2024, 3, 15, 18, 0)
    ]
}).with_columns(
    pl.col("ts").dt.replace_time_zone("UTC")
)

print(df)


## <strong style="color:#5e17eb"> Replace a Time zone </strong>

df.with_columns(
    pl.col("ts").dt.convert_time_zone("Asia/Kolkata").alias("ts_IST")
)


## <strong style="color:#5e17eb">  Date Range</strong>

# Create a date range
dr = pl.date_range(
    start=datetime.datetime(2024, 1, 1),
    end=datetime.datetime(2024, 1, 10),
    interval="3d",
    eager=True
)
dr


df.schema

## <strong style="color:#5e17eb"> Cast  </strong>

df = df.with_columns_seq([
    pl.col("ts").cast(pl.Date).alias("only_date")
])

df = df.with_columns_seq([
    pl.col("only_date").cast(pl.Datetime).alias("back_to_datetime")
])

df

## <strong style="color:#5e17eb"> Summary </strong>

**Goal: Handle local time vs UTC.**

| Sub-Topic                | Function / Method                       | Example          |
| ------------------------ | --------------------------------------- | ---------------- |
| **6.1 Set Timezone**     | `.dt.replace_time_zone("UTC")`          | Convert to UTC   |
| **6.2 Convert Timezone** | `.dt.convert_time_zone("Asia/Kolkata")` | Change time zone |
| **6.3 Remove Timezone**  | `.dt.replace_time_zone(None)`           | Drop tz info     |


# <strong style="color:#5e17eb"> 7. Validation & Cleaning </strong>


user_lf.null_count().collect()

- Rest we already discuss in our previous notebooks

## <strong style="color:#5e17eb"> Summary </strong>

**Goal: Ensure dates/times are valid and clean.**

| Sub-Topic                     | Function / Method                             | Example                                   |
| ----------------------------- | --------------------------------------------- | ----------------------------------------- |
| **7.1 Check Nulls**           | `.is_null()`, `.is_not_null()`                | `df.filter(pl.col("date").is_not_null())` |
| **7.2 Filter by Range**       | `(pl.col("date") >= pl.date(2024,1,1)) & ...` | Keep data in range                        |
| **7.3 Detect Parsing Errors** | `.str.strptime(..., strict=False)`            | Avoid failing on bad dates                |


# <strong style="color:#5e17eb"> 8. Pro Tips for Date/Time in Polars </strong>

- Prefer Native Types: Use `pl.Date` / `pl.Datetime` — they are faster and memory-efficient.

- Always Specify Format: Avoid ambiguous parsing (%Y-%m-%d is faster than guessing).

- Use Lazy API for Big Time Data: Speeds up large range filtering & grouping.

- Use `.start_of()` & `.end_of()` for Grouping: Prevents month/week edge errors.

- Time Arithmetic: .`dt.offset_by()` is safer than adding integers directly.

- Timezones: Convert all to UTC before joins/merges.

- https://docs.pola.rs/api/python/stable/reference/series/temporal.html

<div style="text-align: center;">
  <h4 style="
    display: inline-block;
    color: #5e17eb;
    font-family: 'Segoe UI';
    border-left: 5px solid #5e17eb;
    background-color: #F8F9F9;
    padding: 10px 20px;
    border-radius: 5px;
    text-align: left;
  ">
  <b>
    Thank You 💜
    </b>
  </h4>
</div>
